refactor(gym): 도메인 분리를 위해 폴더로 이동
본문: request와 response폴더 나눠서 저장함
