--- src/main/java/org/helloworld/gymmate/domain/gym/gym/dto/GymResponse.java
@@ -1,4 +1,0 @@
-package org.helloworld.gymmate.domain.gym.gym.dto;
-
-public record GymResponse(Long gymId) {
-}
