+++ src/main/java/org/helloworld/gymmate/domain/gym/gym/dto/response/GymResponse.java
@@ -1,0 +1,4 @@
+package org.helloworld.gymmate.domain.gym.gym.dto.response;
+
+public record GymResponse(Long gymId) {
+}
